Create Spiral Matrix Leet59.java
It is similar to the spiral matrix but here i have took a i variable assigned it to 1 and stored the  i value inside a 2d array arr so that the nsquare numbers i.e. n by n matrix is formed and its values are stored in arr in spiral form.